requirements.txt
# client
openai
tweepy
requests

# caption server
flask
torch
pillow
transformers
